Add user authentication and goal management features with CRUD operations

+++ backend/controllers/goal.controller.js
@@ -1,8 +1,9 @@
 import asyncHandler from "express-async-handler";
-import { Goal } from "../../models/goal.model.js";
+import { Goal } from "../models/goal.model.js";
+import { User } from "../models/user.model.js";
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -15,6 +16,7 @@ const setGoals = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -27,6 +29,20 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error(`user not found`);
+  }
+
+  //make sure logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error(`User not Authorized`);
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -40,6 +56,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error(`user not found`);
+  }
+
+  //make sure logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error(`User not Authorized`);
   }
 
   await Goal.deleteOne(goal);
